fix(umongodb): fix delete file split and stale golden file
- Extract shared deleteOpts/newDeleteCmd into delete.go to satisfy
  one-constructor-per-file rule (rule10)
- Add missing newDeleteSharded with correct action=DeleteUMongoDBShardedCluster
  and idParam=ShardedClusterId
- Remove stale flag=async lines from cmdtree.golden for create-replset
  and create-sharded (flags were removed in prior commit)
