+++ products/umongodb/internal/umongodb/delete_sharded.go
@@ -6,14 +6,14 @@ import (
 	"github.com/ucloud/ucloud-cli/pkg/cli"
 )
 
-// newDeleteReplset implements `umongodb delete-replset`.
-func newDeleteReplset(ctx *cli.Context) *cobra.Command {
+// newDeleteSharded implements `umongodb delete-sharded`.
+func newDeleteSharded(ctx *cli.Context) *cobra.Command {
 	return newDeleteCmd(ctx, deleteOpts{
-		use:        "delete-replset",
-		short:      "Delete MongoDB replica set instances",
-		long:       "Delete one or more MongoDB replica set instances. The cluster is stopped before deletion unless --skip-stop is set.",
-		action:     "DeleteUMongoDBReplSet",
-		idParam:    "ClusterId",
-		idFlagDesc: "Cluster ID(s) of replica set instances to delete.",
+		use:        "delete-sharded",
+		short:      "Delete MongoDB sharded cluster instances",
+		long:       "Delete one or more MongoDB sharded cluster instances. The cluster is stopped before deletion unless --skip-stop is set.",
+		action:     "DeleteUMongoDBShardedCluster",
+		idParam:    "ShardedClusterId",
+		idFlagDesc: "Cluster ID(s) of sharded cluster instances to delete.",
 	})
 }
